+++ build.gradle
@@ -14,8 +14,7 @@ repositories {
 
 dependencies {
     implementation 'org.jsoup:jsoup:1.17.2'
-//    implementation group: 'org.seleniumhq.selenium', name: 'selenium-java', version: '2.41.0'
-//    implementation files('src/libs/darcula.jar')
+    implementation group: 'org.seleniumhq.selenium', name: 'selenium-java', version: '2.41.0'
     implementation group: 'com.opencsv', name: 'opencsv', version: '5.9'
 }
