+++ src/main/java/group17/news_aggregator/news/Blog.java
@@ -1,0 +1,7 @@
+package group17.news_aggregator.news;
+
+public class Blog extends News{
+    public Blog() {
+        setType("Blog");
+    }
+}
