+++ src/main/java/group17/news_aggregator/scraper/Main.java
@@ -2,21 +2,32 @@ package group17.news_aggregator.scraper;
 
 import group17.news_aggregator.csv_converter.CSVConverter;
 import group17.news_aggregator.news.Article;
+import group17.news_aggregator.news.News;
+import group17.news_aggregator.scraper.article.CryptoSlateScraper;
+import group17.news_aggregator.scraper.article.CryptonewsScraper;
+import group17.news_aggregator.scraper.article.CryptopolitanScraper;
+import group17.news_aggregator.scraper.blog.MediumScraper;
 
+import java.util.ArrayList;
 import java.util.List;
 
 public class Main {
     public static void main(String[] args) throws InterruptedException {
         String outputPath = "src/main/resources/data/database.csv";
 
-        CryptopolitanScraper scraper0 = new CryptopolitanScraper();
-        List<Article> list = scraper0.scrapeAll();
+//        List<News> list = new ArrayList<>();
 
-        CryptonewsScraper scraper1 = new CryptonewsScraper();
-        list.addAll(scraper1.scrapeAll());
+//        CryptopolitanScraper scraper0 = new CryptopolitanScraper();
+//        list.addAll(scraper0.scrapeAll());
+//
+//        CryptonewsScraper scraper1 = new CryptonewsScraper();
+//        list.addAll(scraper1.scrapeAll());
+//
+//        CryptoSlateScraper scraper2 = new CryptoSlateScraper();
+//        list.addAll(scraper2.scrapeAll());
 
-        CryptoSlateScraper scraper2 = new CryptoSlateScraper();
-        list.addAll(scraper2.scrapeAll());
+//        MediumScraper scraper3 = new MediumScraper();
+//        list.addAll(scraper3.scrapeAll());
 
         /*
         --------------------------------------
@@ -27,6 +38,12 @@ public class Main {
          */
 
         CSVConverter converter = new CSVConverter();
-        converter.toCSV(list, outputPath, true);
+        List<News> list = converter.fromCSV(outputPath);
+        System.out.println(list.size());
+
+//        MediumScraper scraper3 = new MediumScraper();
+//        list.addAll(scraper3.scrapeAll());
+//
+//        converter.toCSV(list, outputPath, true);
     }
 }
