+++ src/main/resources/group17/news_aggregator/gui/css/medium-reader.css
@@ -1,0 +1,362 @@
+.Button_label__pYjjJ {
+  font-size:16px;
+  line-height:24px;
+  text-transform:uppercase
+}
+@media(max-width:1024px) {
+  .Button_label__pYjjJ {
+    font-size:13px
+  }
+}
+.Button_button__QvmpL {
+  border:none;
+  cursor:pointer;
+  justify-content:center;
+  background-color:transparent;
+  text-decoration:none;
+  display:inline-flex;
+  padding:12px 24px;
+  font-family:inherit;
+  line-height:inherit;
+  font-size:inherit
+}
+.Button_content__rz2Uk {
+  display:flex;
+  align-items:center;
+  justify-content:space-around
+}
+.Button_content__rz2Uk svg {
+  margin-right:calc(var(--base)/2);
+  width:var(--base);
+  height:var(--base)
+}
+.Button_label__pYjjJ {
+  text-align:center;
+  display:flex;
+  align-items:center
+}
+.Button_appearance--primary__aXtFJ {
+  background-color:var(--theme-elevation-1000);
+  color:var(--theme-elevation-0)
+}
+.Button_appearance--secondary__67RnZ {
+  background-color:transparent;
+  box-shadow:inset 0 0 0 1px var(--theme-elevation-1000)
+}
+.Button_primary--invert__abvn4 {
+  background-color:var(--theme-elevation-0);
+  color:var(--theme-elevation-1000)
+}
+.Button_secondary--invert__OpT85 {
+  background-color:var(--theme-elevation-1000);
+  box-shadow:inset 0 0 0 1px var(--theme-elevation-0)
+}
+.Button_appearance--default__Hxo2A,
+.Button_appearance--none__nd_sY {
+  padding:0;
+  color:var(--theme-text)
+}
+.Button_appearance--none__nd_sY .Button_label__pYjjJ {
+  text-transform:none;
+  line-height:inherit;
+  font-size:inherit
+}
+.readmedium-translated-content {
+  display:inline-block;
+  margin-top:8px
+}
+.read-medium-post {
+  -webkit-text-size-adjust:100%;
+  text-rendering:optimizeLegibility;
+  -webkit-font-smoothing:antialiased;
+  word-break:break-word;
+  word-wrap:break-word
+}
+.read-medium-post * {
+  max-width:100%
+}
+.read-medium-post iframe {
+  width:100%;
+  min-height:240px
+}
+.read-medium-post code {
+  word-wrap:break-word;
+  word-break:break-word;
+  line-height:32px;
+  letter-spacing:-.003em;
+  font-weight:400;
+  background-color:#f2f2f2;
+  padding:2px 4px;
+  font-size:75%;
+  font-family:source-code-pro,Menlo,Monaco,Courier New,Courier,monospace
+}
+.read-medium-post code,
+.read-medium-post h1 {
+  color:#242424;
+  font-style:normal;
+  box-sizing:inherit
+}
+.read-medium-post h1 {
+  font-family:sohne,Helvetica Neue,Helvetica,Arial,sans-serif;
+  font-weight:600;
+  font-size:24px;
+  margin:1.95em 0 -.28em;
+  line-height:30px;
+  letter-spacing:-.016em
+}
+.read-medium-post h1:first-child {
+  font-weight:700;
+  font-size:28px;
+  margin:32px 0;
+  line-height:52px;
+  letter-spacing:-.011em
+}
+.read-medium-post h2 {
+  box-sizing:inherit;
+  font-family:sohne,Helvetica Neue,Helvetica,Arial,sans-serif;
+  color:#242424;
+  font-style:normal;
+  font-weight:600;
+  font-size:20px;
+  line-height:24px;
+  letter-spacing:0;
+  margin:1.72em 0
+}
+.read-medium-post p {
+  color:#242424;
+  word-break:break-word;
+  margin:2.14em 0 -.46em
+}
+.read-medium-post a,
+.read-medium-post p {
+  box-sizing:inherit;
+  font-weight:400;
+  font-style:normal;
+  font-family:source-serif-pro,Georgia,Cambria,Times New Roman,Times,serif;
+  font-size:20px;
+  line-height:32px;
+  letter-spacing:-.003em
+}
+.read-medium-post a {
+  -webkit-tap-highlight-color:rgba(0,0,0,0);
+  color:inherit;
+  text-decoration:underline
+}
+.read-medium-post pre {
+  background:#f9f9f9;
+  border-radius:4px;
+  overflow-x:auto;
+  font-family:source-code-pro,Menlo,Monaco,Courier New,Courier,monospace;
+  padding:32px;
+  border:1px solid #e5e5e5;
+  margin:56px 0 0;
+  font-size:14px
+}
+.read-medium-post li,
+.read-medium-post pre {
+  font-weight:400;
+  box-sizing:inherit;
+  color:#242424
+}
+.read-medium-post li {
+  font-style:normal;
+  font-family:source-serif-pro,Georgia,Cambria,Times New Roman,Times,serif;
+  margin-bottom:-.46em;
+  list-style:disc none;
+  margin-left:30px;
+  padding-left:0;
+  font-size:20px;
+  line-height:32px;
+  letter-spacing:-.003em;
+  margin-top:1.14em
+}
+.read-medium-post ul {
+  word-break:break-word;
+  word-wrap:break-word;
+  margin:0;
+  padding:0;
+  list-style:none
+}
+.read-medium-post blockquote,
+.read-medium-post ul {
+  color:rgba(0,0,0,.8);
+  font-family:medium-content-sans-serif-font,-apple-system,BlinkMacSystemFont,Segoe UI,Roboto,Oxygen,Ubuntu,Cantarell,Open Sans,Helvetica Neue,sans-serif;
+  font-weight:400;
+  box-sizing:inherit
+}
+.read-medium-post blockquote {
+  box-shadow:inset 3px 0 0 0 #242424;
+  padding-left:23px;
+  margin:0 0 0 -20px
+}
+.read-medium-post figure {
+  margin-top:56px;
+  padding-top:5px;
+  padding-bottom:5px;
+  clear:both
+}
+.read-medium-post figcaption {
+  font-size:14px;
+  line-height:20px;
+  color:#999;
+  margin-top:8px;
+  text-align:center
+}
+@media screen and (max-width:680px) {
+  .read-medium-post h1 {
+    font-size:20px;
+    line-height:24px
+  }
+  .read-medium-post h1:first-child {
+    font-size:24px;
+    line-height:38px
+  }
+  .read-medium-post h2 {
+    font-size:16px;
+    line-height:20px
+  }
+  .read-medium-post p {
+    margin-top:12px
+  }
+  .read-medium-post a,
+  .read-medium-post li,
+  .read-medium-post p {
+    font-size:18px;
+    line-height:28px
+  }
+  .read-medium-post pre {
+    font-size:16px;
+    line-height:18px
+  }
+}
+html[data-theme=dark] .read-medium-post {
+  color:#fff
+}
+html[data-theme=dark] .read-medium-post code {
+  background:dimgray;
+  color:#fff
+}
+html[data-theme=dark] .read-medium-post blockquote,
+html[data-theme=dark] .read-medium-post h1,
+html[data-theme=dark] .read-medium-post h2,
+html[data-theme=dark] .read-medium-post li,
+html[data-theme=dark] .read-medium-post p {
+  background:#000;
+  color:#fff
+}
+html[data-theme=dark] .read-medium-post pre {
+  background:#999;
+  border:none
+}
+html[data-theme=dark] .read-medium-post figcaption {
+  color:#d3d3d3
+}
+html[data-theme=dark] .read-medium-post blockquote {
+  box-shadow:inset 3px 0 0 0 hsla(0,0%,100%,.8)
+}
+pre code.hljs {
+  display:block;
+  overflow-x:auto;
+  padding:1em
+}
+code.hljs {
+  padding:3px 5px
+}
+/*!
+  Theme: GitHub
+  Description: Light theme as seen on github.com
+  Author: github.com
+  Maintainer: @Hirse
+  Updated: 2021-05-15
+
+  Outdated base version: https://github.com/primer/github-syntax-light
+  Current colors taken from GitHub's CSS
+*/
+.hljs {
+  color:#24292e;
+  background:#fff
+}
+.hljs-doctag,
+.hljs-keyword,
+.hljs-meta .hljs-keyword,
+.hljs-template-tag,
+.hljs-template-variable,
+.hljs-type,
+.hljs-variable.language_ {
+  color:#d73a49
+}
+.hljs-title,
+.hljs-title.class_,
+.hljs-title.class_.inherited__,
+.hljs-title.function_ {
+  color:#6f42c1
+}
+.hljs-attr,
+.hljs-attribute,
+.hljs-literal,
+.hljs-meta,
+.hljs-number,
+.hljs-operator,
+.hljs-selector-attr,
+.hljs-selector-class,
+.hljs-selector-id,
+.hljs-variable {
+  color:#005cc5
+}
+.hljs-meta .hljs-string,
+.hljs-regexp,
+.hljs-string {
+  color:#032f62
+}
+.hljs-built_in,
+.hljs-symbol {
+  color:#e36209
+}
+.hljs-code,
+.hljs-comment,
+.hljs-formula {
+  color:#6a737d
+}
+.hljs-name,
+.hljs-quote,
+.hljs-selector-pseudo,
+.hljs-selector-tag {
+  color:#22863a
+}
+.hljs-subst {
+  color:#24292e
+}
+.hljs-section {
+  color:#005cc5;
+  font-weight:700
+}
+.hljs-bullet {
+  color:#735c0f
+}
+.hljs-emphasis {
+  color:#24292e;
+  font-style:italic
+}
+.hljs-strong {
+  color:#24292e;
+  font-weight:700
+}
+.hljs-addition {
+  color:#22863a;
+  background-color:#f0fff4
+}
+.hljs-deletion {
+  color:#b31d28;
+  background-color:#ffeef0
+}
+.Gutter_gutter__QB0_n {
+  max-width:1920px;
+  margin-left:auto;
+  margin-right:auto
+}
+.Gutter_gutterLeft__9iSai {
+  padding-left:var(--gutter-h)
+}
+.Gutter_gutterRight__4jfEx {
+  padding-right:var(--gutter-h)
+}
